version 1.0

+++ hhSalonAPI/hhSalon.Services/Services/Implementations/ServicesService.cs
@@ -19,6 +19,11 @@ namespace hhSalon.Services.Services.Implementations
         {
             var services = await _context.Services.Include(s => s.ServiceGroup).Where(s => s.ServiceGroup.GroupId == groupId).ToListAsync();
 
+            if( services is null || services.Count == 0)
+            {
+                throw new Exception("Empty");
+            }
+
             List<ServiceVM> servicesVM = new List<ServiceVM>();
 
             foreach (var service in services)
@@ -33,11 +38,19 @@ namespace hhSalon.Services.Services.Implementations
                 servicesVM.Add(serviceVM);
             }
             return servicesVM;
+        
         }
 
 
         public async Task AddNewServiceAsync(ServiceVM newService)
         {
+            int count = _context.Services.Where(s => s.Name == newService.Name).Count();
+
+            if(count > 0)
+            {
+                throw new DbUpdateException("This Service already exists!");
+            }
+
             Service service = new Service
             {
                 Name = newService.Name,
@@ -59,37 +72,49 @@ namespace hhSalon.Services.Services.Implementations
 
         public async Task UpdateServiceAsync(ServiceVM serviceVM)
         {
-            var service = await GetByIdAsync(serviceVM.Id);
-
+            //var service = await GetByIdAsync(serviceVM.Id);
+            var service = _context.Services.Where(s => s.Id == serviceVM.Id).Include(s => s.ServiceGroup).FirstOrDefault();
+            
             if (service != null)
             {
+                
                 service.Name = serviceVM.Name;
                 service.Price = serviceVM.Price;
+
+                if (service.ServiceGroup.GroupId != serviceVM.GroupId)
+                {
+                    var service_group = _context.Services_Groups.Where(sg => sg.ServiceId == service.Id).FirstOrDefault();
+                    _context.Services_Groups.Remove(service_group);
+
+                    ServiceGroup newService_Group = new ServiceGroup()
+                    {
+                        ServiceId = service.Id,
+                        GroupId = serviceVM.GroupId
+                    };
+
+                    await _context.Services_Groups.AddAsync(newService_Group);
+                }
+
+                await _context.SaveChangesAsync();
             }
-
-            var service_group = _context.Services_Groups.Where(sg => sg.ServiceId == service.Id).FirstOrDefault();
-            _context.Services_Groups.Remove(service_group);
-
-            ServiceGroup newService_Group = new ServiceGroup()
-            {
-                ServiceId = service.Id,
-                GroupId = serviceVM.GroupId
-            };
-
-            await _context.Services_Groups.AddAsync(newService_Group);
-            await _context.SaveChangesAsync();
+            else
+                throw new NullReferenceException();           
         }
 
-        public async Task<ServiceVM> GetServiceByIdWithGroupAsync(int id)
+        public async Task<ServiceVM> GetServiceVMByIdAsync(int id)
         {
 
-            //var service = await _context.Services.Where(s => s.Id == id)
-            //	.Include(s => s.Service_Group).ThenInclude(sg => sg.Group).FirstOrDefaultAsync();
-
-            var groupId = await _context.Services_Groups.Where(sg => sg.ServiceId == id).Select(sg => sg.GroupId).FirstOrDefaultAsync();
+			//var service = await _context.Services.Where(s => s.Id == id)
+			//	.Include(s => s.Service_Group).ThenInclude(sg => sg.Group).FirstOrDefaultAsync();
 
 
-            var service = await _context.Services.Where(s => s.Id == id).FirstOrDefaultAsync();
+			var service = await _context.Services.Where(s => s.Id == id).FirstOrDefaultAsync();
+
+            if(service == null)
+                throw new Exception();
+
+			var groupId = await _context.Services_Groups.Where(sg => sg.ServiceId == id).Select(sg => sg.GroupId).FirstOrDefaultAsync();
+                        
 
             return new ServiceVM() { Id = id, Name = service.Name, Price = service.Price, GroupId = groupId };
         }
